reorg components; prevent autoclose on error toasts
