+++ extension/src/components/toasts/errorToast.tsx
@@ -1,4 +1,4 @@
-import { BaseToast, toast } from './BaseToast'
+import { Toast, toast } from './Toast'
 import type { ToastProps } from './ToastProps'
 
 export const errorToast = ({ title, message }: ToastProps) =>
@@ -6,11 +6,9 @@ export const errorToast = ({ title, message }: ToastProps) =>
     ({ dismiss }) => (
       <>
         <div className="flex items-center justify-between gap-4">
-          {title && (
-            <BaseToast.Title className="text-white">{title}</BaseToast.Title>
-          )}
+          {title && <Toast.Title className="text-white">{title}</Toast.Title>}
 
-          <BaseToast.Dismiss
+          <Toast.Dismiss
             className="text-white hover:bg-white/20"
             onDismiss={dismiss}
           />
@@ -19,5 +17,8 @@ export const errorToast = ({ title, message }: ToastProps) =>
         <div className="text-red-50">{message}</div>
       </>
     ),
-    { className: 'border-red-500/60 bg-red-500 dark:border-red-700' },
+    {
+      className: 'border-red-500/60 bg-red-500 dark:border-red-700',
+      autoClose: false,
+    },
   )
